Refactor navigation for slightly better structure
